requirements.txt
Django==2.2.19
pytz==2022.7.1
sqlparse==0.4.3
django-colorfield==0.8.0
